keybindings hook
Add a hook specifically for custom keybindings. The advantage of this over the
startup hook is that we can assume it's safe to reload this one while Sup is
running. The "modes" variable provides a debatably-useful mapping between
user-visible mode names and the actual classes, for use with class_eval.